third_party/WebKit/LayoutTests/bluetooth/service/getCharacteristics/characteristics-not-found-with-uuid.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  let expected = new DOMException(
    'No Characteristics matching UUID ' + battery_level.uuid + ' found in Service with UUID ' +
      heart_rate.uuid + '.',
    'NotFoundError');
  return setBluetoothFakeAdapter('HeartRateAdapter')
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: ['heart_rate']}]}))
    .then(device => device.gatt.connect())
    .then(gattServer => gattServer.getPrimaryService('heart_rate'))
    .then(service => assert_promise_rejects_with_message(
      service.getCharacteristics('battery_level'), expected));
}, 'Request for absent characteristics with UUID. Reject with NotFoundError.');
</script>
